@@ -1,6 +1,7 @@
 import 'dart:io';
 
 import 'package:elmotatawera_events/bussiness_logic/cubit/app_cubit.dart';
+import 'package:elmotatawera_events/data/constant/route_name_manager.dart';
 import 'package:elmotatawera_events/data/constant/size_manager.dart';
 import 'package:elmotatawera_events/presentation/wigets/core/app_text/text_blue.dart';
 import 'package:elmotatawera_events/presentation/wigets/core/widgets/custom_button.dart';
@@ -16,26 +17,37 @@ class EventPage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    AppCubit myCubit=BlocProvider.of(context);
+    AppCubit myCubit = BlocProvider.of(context);
     return SlidingUpPanel(
       parallaxEnabled: true,
       parallaxOffset: 0.5,
       color: Colors.transparent,
-      body: myCubit.selectedEventModel!.img==null? Center(
-        child: Column(mainAxisSize: MainAxisSize.min,
-          children: [
-            GlobalRichText(firstString: "Using",secondString: "El Motatawera",),
-            SizedBox(height: SizeManager.size28,),
-            CustomButton(text: "Attendance",onTap: (){
-
-            },)
-          ],
-        ),
-      ): Image.file(
-        File(myCubit.selectedEventModel!.img!),
-        fit: BoxFit.fill,
-      ),
-
+      body: myCubit.selectedEventModel!.eventData.img == null
+          ? Center(
+              child: Column(
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  GlobalRichText(
+                    firstString: "Using",
+                    secondString: "El Motatawera",
+                  ),
+                  SizedBox(
+                    height: SizeManager.size28,
+                  ),
+                  CustomButton(
+                    text: "Attendance",
+                    onTap: () {
+                      myCubit.getAllAttendence(myCubit.selectedEventModel!.eventData.docId!);
+                      Navigator.pushNamed(context, RouteNameManager.guestAttendanceScreen);
+                    },
+                  )
+                ],
+              ),
+            )
+          : Image.file(
+              File(myCubit.selectedEventModel!.eventData.img!),
+              fit: BoxFit.fill,
+            ),
       panelBuilder: (scroller) => Container(
         padding: EdgeInsets.all(SizeManager.size12),
         height: double.infinity,
@@ -50,7 +62,7 @@ class EventPage extends StatelessWidget {
           crossAxisAlignment: CrossAxisAlignment.stretch,
           children: [
             TextBlue(
-              myCubit.selectedEventModel!.title,
+              myCubit.selectedEventModel!.eventData.title,
               fontSize: SizeManager.size18,
               fontWeight: FontWeight.bold,
             ),
@@ -58,12 +70,14 @@ class EventPage extends StatelessWidget {
               mainAxisAlignment: MainAxisAlignment.spaceBetween,
               children: [
                 TextBlue(
-                  myCubit.selectedEventModel!.isPaid ? "Paid" : "un paid",
+                  myCubit.selectedEventModel!.eventData.isPaid
+                      ? "Paid"
+                      : "un paid",
                   fontWeight: FontWeight.bold,
                   fontSize: SizeManager.size12,
                 ),
                 TextBlue(
-                  "price : ${myCubit.selectedEventModel!.price}",
+                  "price : ${myCubit.selectedEventModel!.eventData.price}",
                   fontWeight: FontWeight.bold,
                   fontSize: SizeManager.size12,
                 ),
@@ -73,42 +87,45 @@ class EventPage extends StatelessWidget {
               mainAxisAlignment: MainAxisAlignment.spaceBetween,
               children: [
                 TextBlue(
-                  "${myCubit.selectedEventModel!.dateTime.year} - ${myCubit.selectedEventModel!.dateTime.month}  - ${myCubit.selectedEventModel!.dateTime.day} ",
+                  "${myCubit.selectedEventModel!.eventData.dateTime.year} - ${myCubit.selectedEventModel!.eventData.dateTime.month}  - ${myCubit.selectedEventModel!.eventData.dateTime.day} ",
                   fontWeight: FontWeight.bold,
                   fontSize: SizeManager.size12,
                 ),
                 TextBlue(
-                  "${myCubit.selectedEventModel!.dateTime.hour}:${myCubit.selectedEventModel!.dateTime.minute} ",
+                  "${myCubit.selectedEventModel!.eventData.dateTime.hour}:${myCubit.selectedEventModel!.eventData.dateTime.minute} ",
                   fontWeight: FontWeight.bold,
                   fontSize: SizeManager.size12,
                 )
               ],
             ),
             TextBlue(
-              "Location : ${myCubit.selectedEventModel!.location}",
+              "Location : ${myCubit.selectedEventModel!.eventData.location}",
               fontWeight: FontWeight.bold,
               fontSize: SizeManager.size11,
             ),
             GlobalPressedText(
               text: "Get Maps",
-              onTap: () async{
-                if (!await launchUrl(Uri.parse(myCubit.selectedEventModel!.locationUrl))) {
-                  throw Exception('Could not launch ${myCubit.selectedEventModel!.locationUrl}');
+              onTap: () async {
+                if (!await launchUrl(
+                    Uri.parse(myCubit.selectedEventModel!.eventData.locationUrl))) {
+                  throw Exception(
+                      'Could not launch ${myCubit.selectedEventModel!.eventData.locationUrl}');
                 }
               },
             ),
             TextBlue(
-               "Description",
+              "Description",
               fontWeight: FontWeight.bold,
               fontSize: SizeManager.size12,
             ),
             TextBlue(
-              myCubit.selectedEventModel!.description ?? "",
+              myCubit.selectedEventModel!.eventData.description ?? "",
               fontWeight: FontWeight.bold,
               fontSize: SizeManager.size10,
             ),
-            TextBlue("People Count: "+
-              myCubit.selectedEventModel!.peopleCount.toString(),
+            TextBlue(
+              "People Count: " +
+                  myCubit.selectedEventModel!.eventData.peopleCount.toString(),
               fontWeight: FontWeight.bold,
               fontSize: SizeManager.size12,
             ),
