@@ -1,15 +1,19 @@
 import 'package:elmotatawera_events/bussiness_logic/cubit/app_cubit.dart';
+import 'package:elmotatawera_events/data/constant/color_manager.dart';
 import 'package:elmotatawera_events/data/constant/route_name_manager.dart';
+import 'package:elmotatawera_events/data/constant/size_manager.dart';
 import 'package:elmotatawera_events/data/model/guest_model.dart';
-import 'package:elmotatawera_events/presentation/wigets/core/app_text/text_blue.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_slidable/flutter_slidable.dart';
+import 'package:google_fonts/google_fonts.dart';
 
 class GuestTile extends StatelessWidget {
   final GuestModel? guestModel;
+  final Function(BuildContext)? onPressedDelete;
+  final Function(BuildContext)? onPressedUpdate;
 
-  const GuestTile({Key? key,required this.guestModel}) : super(key: key);
+  const GuestTile({Key? key, required this.guestModel,this.onPressedDelete,this.onPressedUpdate}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -22,29 +26,51 @@ class GuestTile extends StatelessWidget {
                 .getSelectedQrInvition(guestModel!);
             Navigator.pushNamed(context, RouteNameManager.inviteScreen);
           },
-          title: TextBlue(guestModel!.guestData.firstName +
+          title: Text(guestModel!.guestData.firstName +
               " " +
-              guestModel!.guestData.lastName),
-          subtitle: TextBlue(guestModel!.guestData.peopleCount.toString()),
-          trailing:
-              TextBlue(guestModel!.isConfirmed ? "Confirmed" : "Un Confirmed"),
+              guestModel!.guestData.lastName,
+            style: GoogleFonts.poppins(
+                fontSize: SizeManager.size12,
+                color: ColorManager.deepBlue),),
+          subtitle: Text("Count "+guestModel!.guestData.peopleCount.toString(),
+            style: GoogleFonts.poppins(
+                fontSize: SizeManager.size12,
+                color: ColorManager.deepBlue),),
+          trailing: guestModel!.isConfirmed
+              ? Text(
+                  "Confirmed",
+                  style: GoogleFonts.poppins(
+                      fontSize: SizeManager.size12,
+                      color: ColorManager.greenColor),
+                )
+              : Text(
+                  "Un Confirmed",
+                  style: GoogleFonts.poppins(
+                      fontSize: SizeManager.size12,
+                      color: ColorManager.errorColor),
+                ),
         ),
         startActionPane: ActionPane(
           motion: ScrollMotion(),
           children: [
             SlidableAction(
-              onPressed: (c) {},
-              backgroundColor: Color(0xFFFE4A49),
+              onPressed: (context){
+                BlocProvider.of<AppCubit>(context).deleteGuest(guestModel!);
+              },
+              backgroundColor: ColorManager.errorColor,
               foregroundColor: Colors.white,
               icon: Icons.delete,
               label: 'Delete',
             ),
             SlidableAction(
-              onPressed: (c) {},
+              onPressed: (context){
+                guestModel!.isConfirmed=!guestModel!.isConfirmed;
+                BlocProvider.of<AppCubit>(context).updateGuestData(guestModel!);
+              },
               backgroundColor: Color(0xFF21B7CA),
               foregroundColor: Colors.white,
-              icon: Icons.share,
-              label: 'Share',
+              icon: Icons.update,
+              label:guestModel!.isConfirmed? 'un Confirmed':"Confirmed",
             ),
           ],
         ),
